requirements.txt
alembic==1.7.4
anyio==3.3.4
argon2-cffi==21.1.0
atomicwrites==1.4.1
attrs==21.2.0
autopage==0.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
bottle==0.12.23
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cliff==3.9.0
cmaes==0.8.2
cmd2==2.2.0
colorama==0.4.4
colorlog==6.5.0
cplex==22.1.0.0
cycler==0.11.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
feature-engine==1.1.2
fonttools==4.37.1
gmpy2==2.1.4
greenlet==1.1.2
htmlmin==0.1.12
idna==3.3
ImageHash==4.3.1
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==4.8.1
importlib-resources==5.3.0
iniconfig==1.1.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lazy_loader==0.1rc2
llvmlite==0.39.1
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mdlp-discretization==0.3.3
missingno==0.5.1
mistune==0.8.4
multimethod==1.9
multipledispatch==0.6.0
natsort==8.2.0
nbclassic==0.3.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.8
nose==1.3.7
notebook==6.4.5
numba==0.56.4
numpy==1.21.3
openpyxl==3.0.10
optuna==2.10.0
optuna-dashboard==0.7.3
packaging==21.3
pandas==1.3.4
pandas-flavor==0.3.0
pandas-profiling==3.4.0
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.2
patsy==0.5.2
pbr==5.6.0
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
prettytable==2.2.1
prometheus-client==0.11.0
prompt-toolkit==3.0.21
psycopg2==2.9.5
ptyprocess==0.7.0
py==1.11.0
pycparser==2.20
pydantic==1.10.2
Pygments==2.10.0
pyjanitor==0.22.0
pyparsing==3.0.1
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.4.1
pywin32==304
pywinpty==2.0.8
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-unixsocket==0.2.0
rulelist==0.2.0
SciencePlots==2.0.1
scikit-learn==1.2.0
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
statsmodels==0.13.5
stevedore==3.5.0
tangled-up-in-unicode==0.2.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing==3.7.4.3
typing_extensions==4.4.0
uncertainty-calibration==0.1.4
urllib3==1.26.7
visions==0.7.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
wincertstore==0.2
xarray==2022.11.0
xgboost==1.7.1
zipp==3.6.0
